# By 郑林琴
# Via 郑林琴
* 'master' of https://github.com/Binglin/BaseTableViewController:
  add empty error  pull push category

+++ BaseTableViewController/CommonUIConsts.m
@@ -9,4 +9,4 @@
 #import "CommonUIConsts.h"
 
 
-CGFloat const height_top_navigationBar = 44;
+float const height_top_navigationBar = 44;
